Update Hands-on lab step-by-step - App modernization.md
Contoso, Ltd. and Contoso Insurance are both referenced in this document.  Contoso, Ltd is on the approved fictitious names list, but Contonso Insurance is not, however it is used extensively in the document and appears in the images as well. I left as-is. Let me know if you would like me to change the name. The ppt does not reference Contoso Insurance, but only Contoso, Ltd.